[Refactor] Review - 조회 로직 분리
