+++ src/main/java/com/inq/wishhair/wesharewishhair/review/controller/ReviewController.java
@@ -1,18 +1,13 @@
 package com.inq.wishhair.wesharewishhair.review.controller;
 
 import com.inq.wishhair.wesharewishhair.auth.config.resolver.ExtractPayload;
-import com.inq.wishhair.wesharewishhair.review.controller.dto.response.PagedReviewResponse;
-import com.inq.wishhair.wesharewishhair.review.domain.Review;
 import com.inq.wishhair.wesharewishhair.review.service.ReviewService;
 import com.inq.wishhair.wesharewishhair.review.controller.dto.request.ReviewRequest;
-import com.inq.wishhair.wesharewishhair.review.service.dto.response.ReviewResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.net.URI;
-import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
@@ -30,18 +25,5 @@ public class ReviewController {
         return ResponseEntity
                 .created(URI.create("/api/review/" + reviewId))
                 .build();
-    }
-
-    @GetMapping("/review")
-    public ResponseEntity<PagedReviewResponse> getReviews(Pageable pageable,
-                                                           @RequestParam String condition) {
-
-        List<ReviewResponse> result = reviewService.getReviews(pageable, condition);
-
-        return ResponseEntity.ok(toPagedResponse(result));
-    }
-
-    private PagedReviewResponse toPagedResponse(List<ReviewResponse> result) {
-        return PagedReviewResponse.of(result);
     }
 }
