+++ src/main/java/com/inq/wishhair/wesharewishhair/review/service/ReviewService.java
@@ -51,20 +51,4 @@ public class ReviewService {
         );
         return reviewRepository.save(review).getId();
     }
-
-    public List<ReviewResponse> getReviews(Pageable pageable, String condition) {
-        List<Review> reviews = reviewRepository.findReviewByPaging(pageable);
-        // Query 에서 정렬이 안돼서 Service 에서 정렬
-        if (condition.equals(Condition.LIKES)) {
-            reviews.sort((a, b) -> Integer.compare(b.getLikeReviews().size(), a.getLikeReviews().size()));
-        }
-
-        return toResponse(reviews);
-    }
-
-    private List<ReviewResponse> toResponse(List<Review> reviews) {
-        return reviews.stream()
-                .map(ReviewResponse::new)
-                .toList();
-    }
 }
